Fixed deploy
